feat(eip): model the unsuccessful Forward_Open and enable the driver testsuites
A rejected Forward_Open does not answer in the successful layout with a
non-zero status: CIP replies in a shorter "unsuccessful" format without
connection ids or APIs. CipConnectionManagerResponse is now discriminated on
the general status into Success and Failure variants, so the reply's type says
whether the connection was opened and only the successful one carries a
connection id. Both drivers were updated accordingly, and the probe in
plc4j reads the status off the shared CipServiceResponse supertype instead of
type-testing each concrete response.

plc4go had never received the capability-probe change from #2674, so it still
sent a single Get_Attribute_All while plc4j probed each class with
Get_Attribute_Single. The shared testsuites hid this until their setup steps
were corrected. The Go handshake is now ported, including "keep the state
achieved" on a failed probe.

Also in here, all uncovered by finally being able to run the suites:

- EIPDriverIT is enabled (see the preceding commits for why it could not run)
  and EIPConnectedDriverIT / EIPMessageRouterDriverIT /
  EIPForcedUnconnectedDriverIT are added, so the connection-manager,
  message-router and forced-unconnected handshakes are exercised. Those three
  testsuite files existed but were referenced by no test at all.
- The setup steps describe the actual probe sequence, and the write
  testcases' hand-written implicit lengths are corrected (they were 2 bytes
  short, never validated while the suite was disabled).
- EipTag.getAddressString() is implemented rather than throwing
  NotImplementedException, which the testsuite runner calls when logging.
- A connection-serial-number configuration parameter, which the connected
  testsuite already passed but the driver ignored in favour of a random value,
  making the Forward_Open exchange unreproducible. Default 0 keeps the random
  per-connection behaviour CIP wants.
- A parser-serializer case for a rejected Forward_Open, and the
  Get Attribute Single case moved out of the driver testsuite, where it could
  never run: a driver testcase needs an api-request to make the driver send
  anything.

+++ plc4go/protocols/eip/readwrite/model/CipConnectedResponse.go
@@ -238,7 +238,7 @@ func (m *_CipConnectedResponse) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
 
-func (m *_CipConnectedResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint8, connected bool, serviceLen uint16) (__cipConnectedResponse CipConnectedResponse, err error) {
+func (m *_CipConnectedResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint16, connected bool, serviceLen uint16) (__cipConnectedResponse CipConnectedResponse, err error) {
 	m.CipServiceResponseContract = parent
 	parent._SubType = m
 	positionAware := readBuffer
