+++ plc4go/protocols/eip/readwrite/model/CipReadResponse.go
@@ -238,7 +238,7 @@ func (m *_CipReadResponse) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
 
-func (m *_CipReadResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint8, connected bool, serviceLen uint16) (__cipReadResponse CipReadResponse, err error) {
+func (m *_CipReadResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint16, connected bool, serviceLen uint16) (__cipReadResponse CipReadResponse, err error) {
 	m.CipServiceResponseContract = parent
 	parent._SubType = m
 	positionAware := readBuffer
