+++ plc4go/protocols/eip/readwrite/model/MultipleServiceResponse.go
@@ -264,7 +264,7 @@ func (m *_MultipleServiceResponse) GetLengthInBytes(ctx context.Context) uint64 
 	return m.GetLengthInBits(ctx) / 8
 }
 
-func (m *_MultipleServiceResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint8, connected bool, serviceLen uint16) (__multipleServiceResponse MultipleServiceResponse, err error) {
+func (m *_MultipleServiceResponse) parse(ctx context.Context, readBuffer utils.ReadBuffer, parent *_CipServiceResponse, extStatusSize uint16, connected bool, serviceLen uint16) (__multipleServiceResponse MultipleServiceResponse, err error) {
 	m.CipServiceResponseContract = parent
 	parent._SubType = m
 	positionAware := readBuffer
